resize and drag dashboard items
