+++ frontend/src/layout/Sidebar.js
@@ -22,6 +22,7 @@ import { Link } from 'lib/components/Link'
 import { sceneLogic } from 'scenes/sceneLogic'
 import { dashboardsModel } from '~/models/dashboardsModel'
 import whiteLogo from './_assets/white-logo.svg'
+import { triggerResizeAfterADelay } from 'lib/utils'
 
 const itemStyle = { display: 'flex', alignItems: 'center' }
 
@@ -77,7 +78,7 @@ export default function Sidebar(props) {
     }
 
     return (
-        <Layout.Sider breakpoint="lg" collapsedWidth="0" className="bg-dark">
+        <Layout.Sider breakpoint="lg" collapsedWidth="0" className="bg-dark" onCollapse={triggerResizeAfterADelay}>
             <Menu
                 className="h-100 bg-dark"
                 theme="dark"
