+++ frontend/src/lib/utils.js
@@ -220,3 +220,13 @@ export const idToKey = (array, keyField = 'id') => {
 }
 
 export const delay = ms => new Promise(resolve => window.setTimeout(resolve, ms))
+
+// Trigger a window.reisize event a few times 0...2 sec after the menu was collapsed/expanded
+// We need this so the dashboard resizes itself properly, as the available div width will still
+// change when the sidebar's expansion is animating.
+export function triggerResizeAfterADelay() {
+    const dispatch = () => window.dispatchEvent(new Event('resize'))
+    for (const delay of [0, 50, 100, 300, 500, 700, 1000, 2000]) {
+        window.setTimeout(dispatch, delay)
+    }
+}
