+++ frontend/src/scenes/dashboard/DashboardItems.scss
@@ -4,7 +4,14 @@
     padding: 15px;
     box-shadow: 0 8px 10px rgba(0, 0, 0, 0.1);
     border-radius: 5px;
-    margin-bottom: 30px;
+    overflow: hidden;
+
+    .dashboard-item-container {
+        display: flex;
+        flex-direction: column;
+        width: 100%;
+        height: 100%;
+    }
 
     .dashboard-item-header {
         height: 32px;
@@ -22,6 +29,9 @@
                 font-size: 14px;
                 letter-spacing: 0.5px;
                 color: hsla(210, 26%, 16%, 1);
+                &:hover {
+                    text-decoration: underline;
+                }
             }
         }
         .dashboard-item-settings {
@@ -39,8 +49,7 @@
     }
     .dashboard-item-content {
         overflow: hidden;
-        height: 25vh;
-        max-height: 30vh;
+        flex-grow: 1;
         position: relative;
     }
 
@@ -100,4 +109,63 @@
             stroke: hsla(145, 60%, 50%, 1);
         }
     }
+}
+
+.react-grid-layout {
+    position: relative;
+    transition: height 200ms ease;
+    margin-bottom: 50px;
+}
+.react-grid-item {
+    transition: all 200ms ease;
+    transition-property: left, top;
+}
+.react-grid-item.cssTransforms {
+    transition-property: transform;
+}
+.react-grid-item.resizing {
+    z-index: 1;
+    will-change: width, height;
+}
+
+.react-grid-item.react-draggable-dragging {
+    transition: none;
+    z-index: 3;
+    will-change: transform;
+}
+
+.react-grid-item.dropping {
+    visibility: hidden;
+}
+
+.react-grid-item.react-grid-placeholder {
+    background: red;
+    opacity: 0.2;
+    transition-duration: 100ms;
+    z-index: 2;
+    user-select: none;
+}
+
+.react-grid-item > .react-resizable-handle {
+    position: absolute;
+    width: 20px;
+    height: 20px;
+    bottom: 0;
+    right: 0;
+    cursor: se-resize;
+}
+
+.react-grid-item > .react-resizable-handle::after {
+    content: '';
+    position: absolute;
+    right: 3px;
+    bottom: 3px;
+    width: 5px;
+    height: 5px;
+    border-right: 2px solid rgba(0, 0, 0, 0.4);
+    border-bottom: 2px solid rgba(0, 0, 0, 0.4);
+}
+
+.react-resizable-hide > .react-resizable-handle {
+    display: none;
 }
